adds random enemy sizes
