+++ UI/main_window.py
@@ -445,6 +445,7 @@ class MainWindow(QtWidgets.QMainWindow):
                 # 'randomize-entrances': external_settings[''],
                 'randomize-music': external_settings['Randomize_Music'],
                 'randomize-enemies': external_settings['Randomize_Enemies'],
+                'randomize-enemy-sizes': external_settings['Randomize_Enemy_Sizes'],
                 # 'panel-enemies': True if len([s for s in DAMPE_REWARDS if s not in external_settings['']]) > 0 else False,
                 'shuffle-dungeons': external_settings['Shuffled_Dungeons'],
                 # 'dungeon-items': external_settings[''],
@@ -536,6 +537,7 @@ class MainWindow(QtWidgets.QMainWindow):
             # 'randomize-entrances': self.ui.loadingCheck.isChecked(),
             'randomize-music': self.ui.musicCheck.isChecked(),
             'randomize-enemies': self.ui.enemyCheck.isChecked(),
+            'randomize-enemy-sizes': self.ui.enemySizesCheck.isChecked(),
             # 'panel-enemies': True if len([s for s in DAMPE_REWARDS if s not in self.excluded_checks]) > 0 else False,
             'shuffle-dungeons': self.ui.dungeonsCheck.isChecked(),
             # 'dungeon-items': DUNGEON_ITEM_SETTINGS[self.ui.itemsComboBox.currentIndex()],
