+++ UI/ui_form.py
@@ -287,7 +287,7 @@ class Ui_MainWindow(object):
         self.tab_5.setObjectName(u"tab_5")
         self.groupBox_6 = QGroupBox(self.tab_5)
         self.groupBox_6.setObjectName(u"groupBox_6")
-        self.groupBox_6.setGeometry(QRect(10, 14, 711, 101))
+        self.groupBox_6.setGeometry(QRect(10, 14, 711, 141))
         self.groupBox_6.setAlignment(Qt.AlignHCenter|Qt.AlignTop)
         self.bridgeCheck = QCheckBox(self.groupBox_6)
         self.bridgeCheck.setObjectName(u"bridgeCheck")
@@ -300,38 +300,44 @@ class Ui_MainWindow(object):
         self.musicCheck = QCheckBox(self.groupBox_6)
         self.musicCheck.setObjectName(u"musicCheck")
         self.musicCheck.setEnabled(True)
-        self.musicCheck.setGeometry(QRect(369, 64, 151, 21))
+        self.musicCheck.setGeometry(QRect(369, 104, 151, 21))
         self.musicCheck.setFont(font2)
         self.musicCheck.setCheckable(True)
+        self.kanaletCheck = QCheckBox(self.groupBox_6)
+        self.kanaletCheck.setObjectName(u"kanaletCheck")
+        self.kanaletCheck.setGeometry(QRect(10, 64, 151, 21))
+        self.mazeCheck = QCheckBox(self.groupBox_6)
+        self.mazeCheck.setObjectName(u"mazeCheck")
+        self.mazeCheck.setGeometry(QRect(10, 104, 151, 21))
+        self.mazeCheck.setFont(font2)
+        self.mazeCheck.setCheckable(True)
+        self.chestSizesCheck = QCheckBox(self.groupBox_6)
+        self.chestSizesCheck.setObjectName(u"chestSizesCheck")
+        self.chestSizesCheck.setEnabled(True)
+        self.chestSizesCheck.setGeometry(QRect(189, 64, 151, 21))
+        self.chestSizesCheck.setFont(font2)
+        self.chestSizesCheck.setCheckable(True)
+        self.dungeonsCheck = QCheckBox(self.groupBox_6)
+        self.dungeonsCheck.setObjectName(u"dungeonsCheck")
+        self.dungeonsCheck.setEnabled(True)
+        self.dungeonsCheck.setGeometry(QRect(189, 104, 151, 21))
+        self.dungeonsCheck.setFont(font2)
+        self.dungeonsCheck.setCheckable(True)
         self.enemyCheck = QCheckBox(self.groupBox_6)
         self.enemyCheck.setObjectName(u"enemyCheck")
         self.enemyCheck.setEnabled(True)
         self.enemyCheck.setGeometry(QRect(369, 24, 151, 21))
         self.enemyCheck.setFont(font2)
         self.enemyCheck.setCheckable(True)
-        self.kanaletCheck = QCheckBox(self.groupBox_6)
-        self.kanaletCheck.setObjectName(u"kanaletCheck")
-        self.kanaletCheck.setGeometry(QRect(10, 64, 151, 21))
-        self.mazeCheck = QCheckBox(self.groupBox_6)
-        self.mazeCheck.setObjectName(u"mazeCheck")
-        self.mazeCheck.setGeometry(QRect(189, 64, 151, 21))
-        self.mazeCheck.setFont(font2)
-        self.mazeCheck.setCheckable(True)
-        self.chestSizesCheck = QCheckBox(self.groupBox_6)
-        self.chestSizesCheck.setObjectName(u"chestSizesCheck")
-        self.chestSizesCheck.setEnabled(True)
-        self.chestSizesCheck.setGeometry(QRect(548, 24, 151, 21))
-        self.chestSizesCheck.setFont(font2)
-        self.chestSizesCheck.setCheckable(True)
-        self.dungeonsCheck = QCheckBox(self.groupBox_6)
-        self.dungeonsCheck.setObjectName(u"dungeonsCheck")
-        self.dungeonsCheck.setEnabled(True)
-        self.dungeonsCheck.setGeometry(QRect(548, 64, 151, 21))
-        self.dungeonsCheck.setFont(font2)
-        self.dungeonsCheck.setCheckable(True)
+        self.enemySizesCheck = QCheckBox(self.groupBox_6)
+        self.enemySizesCheck.setObjectName(u"enemySizesCheck")
+        self.enemySizesCheck.setEnabled(True)
+        self.enemySizesCheck.setGeometry(QRect(369, 64, 161, 21))
+        self.enemySizesCheck.setFont(font2)
+        self.enemySizesCheck.setCheckable(True)
         self.groupBox_5 = QGroupBox(self.tab_5)
         self.groupBox_5.setObjectName(u"groupBox_5")
-        self.groupBox_5.setGeometry(QRect(10, 130, 521, 141))
+        self.groupBox_5.setGeometry(QRect(10, 169, 521, 141))
         self.groupBox_5.setAlignment(Qt.AlignHCenter|Qt.AlignTop)
         self.badPetsCheck = QCheckBox(self.groupBox_5)
         self.badPetsCheck.setObjectName(u"badPetsCheck")
@@ -367,7 +373,7 @@ class Ui_MainWindow(object):
         self.bookCheck.setGeometry(QRect(10, 64, 151, 21))
         self.groupBox_7 = QGroupBox(self.tab_5)
         self.groupBox_7.setObjectName(u"groupBox_7")
-        self.groupBox_7.setGeometry(QRect(540, 130, 181, 141))
+        self.groupBox_7.setGeometry(QRect(540, 169, 181, 141))
         self.groupBox_7.setAlignment(Qt.AlignHCenter|Qt.AlignTop)
         self.songsCheck = QCheckBox(self.groupBox_7)
         self.songsCheck.setObjectName(u"songsCheck")
@@ -927,13 +933,6 @@ class Ui_MainWindow(object):
 #endif // QT_CONFIG(whatsthis)
         self.musicCheck.setText(QCoreApplication.translate("MainWindow", u"Randomize Music", None))
 #if QT_CONFIG(tooltip)
-        self.enemyCheck.setToolTip("")
-#endif // QT_CONFIG(tooltip)
-#if QT_CONFIG(whatsthis)
-        self.enemyCheck.setWhatsThis(QCoreApplication.translate("MainWindow", u"Randomizes enemies that do not affect logic. Needed kills are left vanilla", None))
-#endif // QT_CONFIG(whatsthis)
-        self.enemyCheck.setText(QCoreApplication.translate("MainWindow", u"Randomize Enemies", None))
-#if QT_CONFIG(tooltip)
         self.kanaletCheck.setToolTip("")
 #endif // QT_CONFIG(tooltip)
 #if QT_CONFIG(whatsthis)
@@ -961,6 +960,20 @@ class Ui_MainWindow(object):
         self.dungeonsCheck.setWhatsThis(QCoreApplication.translate("MainWindow", u"Randomizes Dungeons<br></br><u>Only front entrances are randomized,</u> and only handles non-glitched logics", None))
 #endif // QT_CONFIG(whatsthis)
         self.dungeonsCheck.setText(QCoreApplication.translate("MainWindow", u"Shuffled Dungeons", None))
+#if QT_CONFIG(tooltip)
+        self.enemyCheck.setToolTip("")
+#endif // QT_CONFIG(tooltip)
+#if QT_CONFIG(whatsthis)
+        self.enemyCheck.setWhatsThis(QCoreApplication.translate("MainWindow", u"Randomizes enemies. Required kills are left vanilla", None))
+#endif // QT_CONFIG(whatsthis)
+        self.enemyCheck.setText(QCoreApplication.translate("MainWindow", u"Randomize Enemies", None))
+#if QT_CONFIG(tooltip)
+        self.enemySizesCheck.setToolTip("")
+#endif // QT_CONFIG(tooltip)
+#if QT_CONFIG(whatsthis)
+        self.enemySizesCheck.setWhatsThis(QCoreApplication.translate("MainWindow", u"Randomizes the size of the enemies between 50% and 150%", None))
+#endif // QT_CONFIG(whatsthis)
+        self.enemySizesCheck.setText(QCoreApplication.translate("MainWindow", u"Randomize Enemy Sizes", None))
         self.groupBox_5.setTitle(QCoreApplication.translate("MainWindow", u"Logic Options", None))
 #if QT_CONFIG(tooltip)
         self.badPetsCheck.setToolTip("")
